Better structure directory.
