+++ src/Field.php
@@ -10,9 +10,9 @@ use Yii\Extension\Form\Attribute\GlobalAttributes;
 use Yii\Extension\Form\Attribute\InputAttributes;
 use Yii\Extension\Form\Attribute\WidgetAttributes;
 use Yii\Extension\Form\Contract\PlaceholderContract;
-use Yii\Extension\Form\FieldPart\Error;
-use Yii\Extension\Form\FieldPart\Hint;
-use Yii\Extension\Form\FieldPart\Label;
+use Yii\Extension\Form\Part\Error;
+use Yii\Extension\Form\Part\Hint;
+use Yii\Extension\Form\Part\Label;
 use Yii\Extension\FormModel\Contract\FormModelContract;
 use Yiisoft\Arrays\ArrayHelper;
 use Yiisoft\Definitions\Exception\CircularReferenceException;
